# api/openapi/clob.yaml
openapi: "3.0.3"
info:
  title: Poly Paper Trading - CLOB API
  version: "1.0.0"
  description: |
    Polymarket-compatible CLOB API for paper trading.
    All paths are relative to the /clob mount point.
    Bots set host to http://localhost:8080/clob.
paths:
...
  /auth/api-key:
    post:
      operationId: createApiKey
      tags: [auth]
      summary: Create a new API key (L1 auth)
      parameters:
        - $ref: "#/components/parameters/PolyAddress"
        - $ref: "#/components/parameters/PolySignature"
        - $ref: "#/components/parameters/PolyTimestamp"
        - $ref: "#/components/parameters/PolyNonce"
      responses:
        "200":
          description: API key created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiCreds"
        "401":
          description: Unauthorized
    delete:
      operationId: deleteApiKey
      tags: [auth]
      summary: Delete an API key (L2 auth)
      parameters:
        - $ref: "#/components/parameters/PolyAddress"
        - $ref: "#/components/parameters/PolySignatureL2"
        - $ref: "#/components/parameters/PolyTimestamp"
        - $ref: "#/components/parameters/PolyApiKey"
        - $ref: "#/components/parameters/PolyPassphrase"
      responses:
        "200":
          description: API key deleted
        "401":
          description: Unauthorized
  /auth/derive-api-key:
    get:
      operationId: deriveApiKey
      tags: [auth]
      summary: Derive (retrieve) existing API key (L1 auth)
      parameters:
        - $ref: "#/components/parameters/PolyAddress"
        - $ref: "#/components/parameters/PolySignature"
        - $ref: "#/components/parameters/PolyTimestamp"
        - $ref: "#/components/parameters/PolyNonce"
      responses:
        "200":
          description: API key derived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiCreds"
        "401":
          description: Unauthorized
  /auth/api-keys:
    get:
      operationId: getApiKeys
      tags: [auth]
      summary: List all API keys (L2 auth)
      parameters:
        - $ref: "#/components/parameters/PolyAddress"
        - $ref: "#/components/parameters/PolySignatureL2"
        - $ref: "#/components/parameters/PolyTimestamp"
        - $ref: "#/components/parameters/PolyApiKey"
        - $ref: "#/components/parameters/PolyPassphrase"
      responses:
        "200":
          description: API keys list
          content:
            application/json:
              schema:
                type: object
                properties:
                  apiKeys:
                    type: array
                    items:
                      $ref: "#/components/schemas/ApiCreds"

  # --- Trading (L2) ---
...
  /data/orders:
    get:
      operationId: getOrders
      tags: [query]
      summary: Get user orders (cursor-paginated)
      parameters:
        - name: market
          in: query
          schema:
            type: string
        - name: asset_id
          in: query
          schema:
            type: string
        - name: next_cursor
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/PolyAddress"
        - $ref: "#/components/parameters/PolySignatureL2"
        - $ref: "#/components/parameters/PolyTimestamp"
        - $ref: "#/components/parameters/PolyApiKey"
        - $ref: "#/components/parameters/PolyPassphrase"
      responses:
        "200":
          description: Paginated orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedOrders"
  /data/trades:
    get:
      operationId: getTrades
      tags: [query]
      summary: Get user trades (cursor-paginated)
      parameters:
        - name: market
          in: query
          schema:
            type: string
        - name: asset_id
          in: query
          schema:
            type: string
        - name: next_cursor
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/PolyAddress"
        - $ref: "#/components/parameters/PolySignatureL2"
        - $ref: "#/components/parameters/PolyTimestamp"
        - $ref: "#/components/parameters/PolyApiKey"
        - $ref: "#/components/parameters/PolyPassphrase"
      responses:
        "200":
          description: Paginated trades
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedTrades"
...
components:
  parameters:
    PolyAddress:
      name: POLY_ADDRESS
      in: header
      required: true
      schema:
        type: string
    PolySignature:
      name: POLY_SIGNATURE
      in: header
      required: true
      schema:
        type: string
      description: EIP-712 signature (L1)
    PolySignatureL2:
      name: POLY_SIGNATURE
      in: header
      required: true
      schema:
        type: string
      description: HMAC-SHA256 signature (L2)
    PolyTimestamp:
      name: POLY_TIMESTAMP
      in: header
      required: true
      schema:
        type: string
    PolyNonce:
      name: POLY_NONCE
      in: header
      schema:
        type: string
        default: "0"
    PolyApiKey:
      name: POLY_API_KEY
      in: header
      required: true
      schema:
        type: string
    PolyPassphrase:
      name: POLY_PASSPHRASE
      in: header
      required: true
      schema:
        type: string
  schemas:
    ApiCreds:
      type: object
      properties:
        apiKey:
          type: string
        secret:
          type: string
        passphrase:
          type: string
...
    OpenOrder:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
        owner:
          type: string
        maker_address:
          type: string
        market:
          type: string
        asset_id:
          type: string
        side:
          type: string
        original_size:
          type: string
        size_matched:
          type: string
        price:
          type: string
        associate_trades:
          type: array
          items:
            type: string
        outcome:
          type: string
        created_at:
          type: string
        expiration:
          type: string
        order_type:
          type: string
    TradeItem:
      type: object
      properties:
        id:
          type: string
        taker_order_id:
          type: string
        market:
          type: string
        asset_id:
          type: string
        side:
          type: string
        size:
          type: string
        fee_rate_bps:
          type: string
        price:
          type: string
        status:
          type: string
        match_time:
          type: string
        last_update:
          type: string
        outcome:
          type: string
        owner:
          type: string
        maker_address:
          type: string
        bucket_index:
          type: integer
        transaction_hash:
          type: string
        trader_side:
          type: string
        maker_orders:
          type: array
          items:
            type: object
    PaginatedOrders:
      type: object
      properties:
        next_cursor:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/OpenOrder"
    PaginatedTrades:
      type: object
      properties:
        next_cursor:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/TradeItem"
...